+++ packages/@expo/config-plugins/src/ios/Locales.ts
@@ -1,17 +1,12 @@
 import { ExpoConfig } from '@expo/config-types';
-import JsonFile from '@expo/json-file';
 import fs from 'fs';
-import { join, relative } from 'path';
+import path from 'path';
 import { XcodeProject } from 'xcode';
 
 import { ConfigPlugin } from '../Plugin.types';
 import { addResourceFileToGroup, ensureGroupRecursively, getProjectName } from './utils/Xcodeproj';
 import { withXcodeProject } from '../plugins/ios-plugins';
-import { addWarningIOS } from '../utils/warnings';
-
-type LocaleJson = Record<string, string>;
-type ResolvedLocalesJson = Record<string, LocaleJson>;
-type ExpoConfigLocales = NonNullable<ExpoConfig['locales']>;
+import { getResolvedLocalesAsync, LocaleJson } from '../utils/locales';
 
 export const withLocales: ConfigPlugin = (config) => {
   return withXcodeProject(config, async (config) => {
@@ -38,20 +33,20 @@ export async function setLocalesAsync(
     return project;
   }
   // possibly validate CFBundleAllowMixedLocalizations is enabled
-  const localesMap = await getResolvedLocalesAsync(projectRoot, locales);
+  const localesMap = await getResolvedLocalesAsync(projectRoot, locales, 'ios');
 
   const projectName = getProjectName(projectRoot);
-  const supportingDirectory = join(projectRoot, 'ios', projectName, 'Supporting');
+  const supportingDirectory = path.join(projectRoot, 'ios', projectName, 'Supporting');
 
   // TODO: Should we delete all before running? Revisit after we land on a lock file.
   const stringName = 'InfoPlist.strings';
 
   for (const [lang, localizationObj] of Object.entries(localesMap)) {
-    const dir = join(supportingDirectory, `${lang}.lproj`);
+    const dir = path.join(supportingDirectory, `${lang}.lproj`);
     // await fs.ensureDir(dir);
     await fs.promises.mkdir(dir, { recursive: true });
 
-    const strings = join(dir, stringName);
+    const strings = path.join(dir, stringName);
     const buffer = [];
     for (const [plistKey, localVersion] of Object.entries(localizationObj)) {
       buffer.push(`${plistKey} = "${localVersion}";`);
@@ -67,7 +62,7 @@ export async function setLocalesAsync(
     if (!group?.children.some(({ comment }) => comment === stringName)) {
       // Only write the file if it doesn't already exist.
       project = addResourceFileToGroup({
-        filepath: relative(supportingDirectory, strings),
+        filepath: path.relative(supportingDirectory, strings),
         groupName,
         project,
         isBuildFile: true,
@@ -77,31 +72,4 @@ export async function setLocalesAsync(
   }
 
   return project;
-}
-
-export async function getResolvedLocalesAsync(
-  projectRoot: string,
-  input: ExpoConfigLocales
-): Promise<ResolvedLocalesJson> {
-  const locales: ResolvedLocalesJson = {};
-  for (const [lang, localeJsonPath] of Object.entries(input)) {
-    if (typeof localeJsonPath === 'string') {
-      try {
-        locales[lang] = await JsonFile.readAsync(join(projectRoot, localeJsonPath));
-      } catch {
-        // Add a warning when a json file cannot be parsed.
-        addWarningIOS(
-          `locales.${lang}`,
-          `Failed to parse JSON of locale file for language: ${lang}`,
-          'https://docs.expo.dev/distribution/app-stores/#localizing-your-ios-app'
-        );
-      }
-    } else {
-      // In the off chance that someone defined the locales json in the config, pass it directly to the object.
-      // We do this to make the types more elegant.
-      locales[lang] = localeJsonPath;
-    }
-  }
-
-  return locales;
 }